+++ apps/vue-mri-ui-lib/src/components/helpers/bookmarkItems.ts
